fix: resolve purchase and reception order logic gaps
- Fix missing return in PO cancel controller error handling
- Refactor PO cancel to use validateTransition instead of duplicated check
- Create SubmitPurchaseOrderRequest for PO submit action
- Fix RO update to only set uncompleted status when line items change
- Remove unused permissions (PO delete, PO restore, RO delete)
- Remove unused POActionButtons component
- Change float columns to decimal for monetary values in migrations
- Add decimal casts to PO, POProduct, and ROProduct models
- Remove proof_of_payment fields from PO store request (moved to paid transition)

+++ database/migrations/2024_05_26_200610_create_purchase_order_product.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('purchase_order_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_variant_id')->constrained()->onDelete('cascade');
-            $table->float('quantity', 8);
-            $table->float('price', 8);
-            $table->float('total', 8);
+            $table->decimal('quantity', 10, 4);
+            $table->decimal('price', 10, 2);
+            $table->decimal('total', 10, 2);
             $table->timestamps();
         });
     }
